requirements.txt
# Function dependencies, for example:
# package>=version
beautifulsoup4==4.8.2
requests==2.22.0
lxml==4.5.0
google-cloud-bigquery==2.6.1
google-cloud-secret-manager==2.1.0